moved buildSettingsTree to utils directory
